+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+* Scales have been reworked and are now nearly identical to the [types defined in ChartJS](https://github.com/chartjs/Chart.js/blob/v4.4.3/src/types/index.d.ts)
+* The use of ``Color`` for various coloring related options is no longer required. 
+  * It's now possible to use e.g. strings
+
 ## 1.5.0
 * Added ``JsonIgnore`` to certain fields to help prevent infinite loops #174
 * Corrected model for ``Title`` and ``Font`` #175